Removed a Dev feature unintended for this release
This feature required certain code to be commented out in order to
effectively use the Current Test scene as an example. As I would like
the afformentioned scene to just work without needing to change
anything, I commented out the code myself and edited the scene so others
dont need to.